+++ packages/navigation/atlassian-switcher/src/i18n/index.ts
@@ -28,3 +28,7 @@ export { default as ru } from './ru';
 export { default as sk } from './sk';
 export { default as es } from './es';
 export { default as sv } from './sv';
+export { default as th } from './th';
+export { default as tr } from './tr';
+export { default as uk } from './uk';
+export { default as vi } from './vi';