+++ packages/navigation/atlassian-switcher/src/types.ts
@@ -60,6 +60,8 @@ export enum Feature {
   isDiscoverMoreForEveryoneEnabled = 'isDiscoverMoreForEveryoneEnabled',
   // EMCEE stands for Embedded Marketplace with in the product
   isEmceeLinkEnabled = 'isEmceeLinkEnabled',
+  // Enable Discover section - group suggested product links in Discover section
+  isDiscoverSectionEnabled = 'isDiscoverSectionEnabled',
 }
 
 export type FeatureFlagProps = {
@@ -73,6 +75,8 @@ export type FeatureFlagProps = {
   isDiscoverMoreForEveryoneEnabled?: boolean;
   // Enable Embedded Marketplace within the product.
   isEmceeLinkEnabled?: boolean;
+  // Enable Discover section - group suggested product links in Discover section
+  isDiscoverSectionEnabled?: boolean;
 };
 
 export type FeatureMap = { [key in Feature]: boolean };